MyApps-16: Move cypher to viewmodel submodule.
