--- app/src/main/java/xyz/torquato/myapps/ui/component/cypher/model/CypherUiState.kt
@@ -1,5 +1,0 @@
-package xyz.torquato.myapps.ui.component.cypher.model
-
-data class CypherUiState(
-    val message: String,
-)
