+++ presentation/viewmodel/src/main/java/xyz/torquato/myapps/presentation/viewmodel/cypher/model/CypherUiState.kt
@@ -1,0 +1,5 @@
+package xyz.torquato.myapps.presentation.viewmodel.model
+
+data class CypherUiState(
+    val message: String,
+)
